requirements.txt
langchain
langchain-community
langchain-classic
langchain_groq
langchain-text-splitters
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface
